Auto stash before rebase of "thep2p/cratify"
