+++ src/core/mod.rs
@@ -5,8 +5,8 @@ pub mod testutil;
 
 pub use crate::core::lookup::array_lookup_table::ArrayLookupTable;
 pub use crate::core::lookup::array_lookup_table::LOOKUP_TABLE_LEVELS;
-pub use crate::core::lookup::lookup_table::LookupTable;
-pub use crate::core::lookup::lookup_table::LookupTableLevel;
+pub use crate::core::lookup::LookupTable;
+pub use crate::core::lookup::LookupTableLevel;
 pub use crate::core::model::address::Address;
 pub use crate::core::model::identifier::Identifier;
 pub use crate::core::model::memvec::MembershipVector;
